[engraving] moved some read methods to TRead class
