+++ src/engraving/CMakeLists.txt
@@ -103,14 +103,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/rw/400/restrw.h
     ${CMAKE_CURRENT_LIST_DIR}/rw/400/measurerepeatrw.cpp
     ${CMAKE_CURRENT_LIST_DIR}/rw/400/measurerepeatrw.h
-    ${CMAKE_CURRENT_LIST_DIR}/rw/400/clefrw.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/rw/400/clefrw.h
     ${CMAKE_CURRENT_LIST_DIR}/rw/400/timesigrw.cpp
     ${CMAKE_CURRENT_LIST_DIR}/rw/400/timesigrw.h
-    ${CMAKE_CURRENT_LIST_DIR}/rw/400/groupsrw.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/rw/400/groupsrw.h
-    ${CMAKE_CURRENT_LIST_DIR}/rw/400/keysigrw.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/rw/400/keysigrw.h
     ${CMAKE_CURRENT_LIST_DIR}/rw/400/tread.cpp
     ${CMAKE_CURRENT_LIST_DIR}/rw/400/tread.h
