+++ src/engraving/libmscore/instrchange.cpp
@@ -24,6 +24,7 @@
 
 #include "translation.h"
 #include "rw/xml.h"
+#include "rw/400/tread.h"
 
 #include "keysig.h"
 #include "measure.h"
@@ -206,36 +207,9 @@ void InstrumentChange::write(XmlWriter& xml) const
     xml.endElement();
 }
 
-//---------------------------------------------------------
-//   read
-//---------------------------------------------------------
-
 void InstrumentChange::read(XmlReader& e)
 {
-    while (e.readNextStartElement()) {
-        const AsciiStringView tag(e.name());
-        if (tag == "Instrument") {
-            _instrument->read(e, part());
-        } else if (tag == "init") {
-            _init = e.readBool();
-        } else if (!TextBase::readProperties(e)) {
-            e.unknown();
-        }
-    }
-    if (score()->mscVersion() < 206) {
-        // previous versions did not honor transposition of instrument change
-        // except in ways that it should not have
-        // notes entered before the instrument change was added would not be altered,
-        // so original transposition remained in effect
-        // notes added afterwards would be transposed by both intervals, resulting in tpc corruption
-        // here we set the instrument change to inherit the staff transposition to emulate previous versions
-        // in Note::read(), we attempt to fix the tpc corruption
-        // There is also code in read206 to try to deal with this, but it is out of date and therefore disabled
-        // What this means is, scores created in 2.1 or later should be fine, scores created in 2.0 maybe not so much
-
-        Interval v = staff() ? staff()->part()->instrument(tick())->transpose() : 0;
-        _instrument->setTranspose(v);
-    }
+    rw400::TRead::read(this, e, *e.context());
 }
 
 //---------------------------------------------------------
