+++ src/engraving/rw/400/tread.h
@@ -76,7 +76,10 @@ class Clef;
 class Fingering;
 class Glissando;
 class GradualTempoChange;
+class Groups;
+class Hairpin;
 class Hook;
+class KeySig;
 class Lyrics;
 class MeasureBase;
 class Note;
@@ -112,7 +115,6 @@ public:
     static void read(SystemText* t, XmlReader& xml, ReadContext& ctx);
     static void read(PlayTechAnnotation* a, XmlReader& xml, ReadContext& ctx);
     static void read(RehearsalMark* m, XmlReader& xml, ReadContext& ctx);
-    static void read(InstrumentChange* c, XmlReader& xml, ReadContext& ctx);
     static void read(StaffState* s, XmlReader& xml, ReadContext& ctx);
     static void read(Image* i, XmlReader& xml, ReadContext& ctx);
     static void read(Tuplet* t, XmlReader& xml, ReadContext& ctx);
@@ -149,7 +151,11 @@ public:
     static void read(Fingering* f, XmlReader& xml, ReadContext& ctx);
     static void read(Glissando* g, XmlReader& xml, ReadContext& ctx);
     static void read(GradualTempoChange* c, XmlReader& xml, ReadContext& ctx);
+    static void read(Groups* g, XmlReader& xml, ReadContext& ctx);
+    static void read(Hairpin* h, XmlReader& xml, ReadContext& ctx);
     static void read(Hook* h, XmlReader& xml, ReadContext& ctx);
+    static void read(InstrumentChange* c, XmlReader& xml, ReadContext& ctx);
+    static void read(KeySig* s, XmlReader& xml, ReadContext& ctx);
     static void read(Lyrics* l, XmlReader& xml, ReadContext& ctx);
     static void read(Note* n, XmlReader& xml, ReadContext& ctx);
     static void read(NoteDot* d, XmlReader& xml, ReadContext& ctx);
